pyproject.toml
[tool.poetry]
name = "openwastemap"
version = "0.4"
description = ""
authors = ["Benito Buchheim <benito@openwastemap.org>"]

[tool.poetry.dependencies]
python = "^3.7, <3.11"
Django = "^3.2.8"
python-dotenv = "^0.19.1"
django-sass = "^1.1.0"
mysqlclient = "^2.0.3"
Pillow = "^8.4.0"
numpy = "^1.21.3"
django-email-verification = {git = "https://github.com/AbdoMahfoz/django-email-verification"}

[tool.poetry.dev-dependencies]
pre-commit = "^2.15.0"
black = {version = "^21.10b0", allow-prereleases = true}
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
